+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added all project dependencies from `pubspec.yaml`.
 - Integrated `GetIt` for dependency injection.
 - Added `DebugScreen` widget for testing and debugging purposes.
+- Added application router with `GoRouter` and defined route paths for navigation.
 
 ### Changed
 - Added `public_member_api_docs` rule to `analysis_options.yaml`.